Fixed problem with sticky footer
The sticky footer in this solution caused issues
with flexbox as their solution was written. When
the viewport was shrunk and flex wrap set to wrap
the overflow of items would hide behind the footer.
I converted the body to a flex container and set the
inner container to flex: 1 to push the footer down
but also stop the content from hiding behind the
fixed position of the footer and converted the height
of the footer to padding.